+++ db/migrate/20200226211925_add_index_to_tasks_status.rb
@@ -1,0 +1,5 @@
+class AddIndexToTasksStatus < ActiveRecord::Migration[6.0]
+  def change
+    add_index :tasks, :status
+  end
+end